Progress so far on the user homepage
IF the logged-in user has their information saved in the database, they are redirected to their profile page after signing in. The profile page consists of their data, classes they are enrolled in, and their pending and posted reviews. It also has a link to edit the personal information, so it suffices both the requirement of retrieving and updating info from and the database.