Don't merge this branch. WIP
I tried to include the add class features in Admin's profile page. As of this update, we can successfully add classes, but since ClassList and ClassReviews collections have an overlapping use, creating default reviews for the newly created class might require lots of changes in the already built code. I also edited the default data and collection scheme for ClassList as I encounter multiple errors due to the numeric prefix in the keys ("1star" is changed to "oneStar" and so on). This is still a WIP. Don't merge this brach into masters